web-app/requirements.txt
asgiref==3.2.7
Django==3.1.8
Pillow==8.1.1
psycopg2==2.8.5
pytz==2019.3
